reverts the cron to an API call. Updated readme for support.
