+++ .env.sample
@@ -17,7 +17,7 @@ POSTGRES_URL_NON_POOLING="postgresql://dc:OpenThrone@192.168.0.1/OpenThrone"
 NEXTAUTH_URL=http://localhost:3000
 NEXT_PUBLIC_URL_ROOT=http://localhost:3000
 
-# A few other secret keys you might want to change
+# CHANGE THIS TO A RANDOM/SECURE STRING
 TASK_SECRET="TESTING"
 REVALIDATION_SECRET="CHANGE_ME_TO_REVALIDATE_STATS"
